% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_bill_info.R
\name{cam_bill_info}
\alias{cam_bill_info}
\title{Downloads details of a specific bill by providing id of a bill}
\usage{
cam_bill_info(id_bill, ascii = T)
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads details of a specific bill by providing id of a bill
}
\note{
Requesting data from a long period of time with \code{details = TRUE} will
return a large object in terms of memory. It will also be rather unwieldy, with
many columns.
}
\examples{
cam_bill_info(14784)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
