% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_bill_info_tny.R
\name{cam_bill_info_tny}
\alias{cam_bill_info_tny}
\title{Downloads details of a specific bill by providing type, number and year}
\usage{
cam_bill_info_tny(type, number, year)
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\note{
Requesting data from a long period of time with \code{details = TRUE} will
return a large object in terms of memory. It will also be rather unwieldy, with
many columns.
}
\examples{
cam_bill_info_tny(type="PL", number="3962", year="2008")
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
