% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_plenary_bills.R
\name{cam_plrenary_bills}
\alias{cam_plrenary_bills}
\title{This function lists every bill voted on in plenary.}
\usage{
cam_plrenary_bills(type = "", year)
}
\arguments{
\item{initial_data}{(\code{character}) start date of the period requested.
This parameter must be in the format YYYYMMDD (Year-Month-Day). A value for
this parameter is necessary, all others are optional.}

\item{end_data}{(\code{character}) final date for period requested. Format
YYYYMMDD.}

\item{house}{(\code{character}). The acronym for the legislative house
for which results are requested. Options are SF (\emph{Senado Federal}, Federal Senate)
and CN (\emph{Congresso Nacional}, National Congress - joint meeting of
the Senate and Chamber).}

\item{colegiado}{To Do}

\item{legislator}{To Do}

\item{details}{(\code{logical}). If details is equal to TRUE, the data returned
is an expanded dataset with additional details. This is not recommended unless
necessary.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\note{
Requesting data from a long period of time with \code{details = TRUE} will
return a large object in terms of memory. It will also be rather unwieldy, with
many columns.
}
\examples{
cam_plenary_bills(year=2008)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
