% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_typeauthors_bills.R
\name{cam_typeauthors_bills}
\alias{cam_typeauthors_bills}
\title{Downloads types of authors for bills}
\usage{
cam_typeauthors_bills()
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\note{
Requesting data from a long period of time with \code{details = TRUE} will
return a large object in terms of memory. It will also be rather unwieldy, with
many columns.
}
\examples{
cam_typeauthors_bills()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
