% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cham_bill_info_id.R
\name{cham_bill_info_id}
\alias{cham_bill_info_id}
\title{Downloads details of a specific bill by providing id of a bill}
\usage{
cham_bill_info_id(id_bill, ascii = T)
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads details of a specific bill by providing id of a bill
}
\note{
Requesting data from a long period of time with \code{details = TRUE} will
return a large object in terms of memory. It will also be rather unwieldy, with
many columns.
}
\examples{
cham_bill_info_id(14784)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
