% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cham_bill_list_authors.R
\name{cham_bill_list_authors}
\alias{cham_bill_list_authors}
\title{Downloads list of authors of a specific bill by providing id of a bill}
\usage{
cham_bill_list_authors(bill_id)
}
\arguments{
\item{bill_id}{\code{integer}. The id number of the bill.}
}
\value{
A data-frame.
}
\description{
Downloads list of authors of a specific bill by providing id of a bill
}
\examples{
\donttest{
cham_bill_list_authors(321182)
}
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
