% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cham_deputy_list.R
\name{cham_legislator_list}
\alias{cham_legislator_list}
\title{Downloads and tidies information on the legislators in the Chamber of Deputies}
\usage{
cham_legislator_list()
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads and tidies information on the legislators in the Federal Senate.
}
\examples{
\dontrun{all <- cham_legislator_list()}

}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
