% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cham_typeauthors_bills.R
\name{cham_typeauthors_bills}
\alias{cham_typeauthors_bills}
\title{Downloads types of authors for bills}
\usage{
cham_typeauthors_bills(ascii = T)
}
\arguments{
\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads types of authors for bills
}
\note{
Requesting data from a long period of time with \code{details = TRUE} will
return a large object in terms of memory. It will also be rather unwieldy, with
many columns.
}
\examples{
cham_typeauthors_bills()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
