% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cham_votes.R
\name{cham_votes_year}
\alias{cham_votes_year}
\title{Returns voting information from the Chamber floor for the year
requested}
\usage{
cham_votes_year(year)
}
\arguments{
\item{year}{\code{character} or \code{integer}. Format YYYY}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Returns voting information from the Chamber floor for the year
requested.
}
\examples{
\donttest{
cham_votes_year("2013")
}

}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
