% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_commissions.R
\docType{data}
\name{commissions}
\alias{commissions}
\title{A dataframe of commissions in the Brazilian Federal Senate.}
\format{A data frame with 21 rows and 2 variables}
\usage{
commissions
}
\description{
\itemize{
  \item Commission \code{character}. Full name of the commission.
  \item Name \code{character}. Abbreviation. For use as an argument to
  various congressbr functions (such as \code{sen_senator_comms()}).
}
}
\keyword{datasets}

