% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congressbr-package.r
\docType{package}
\name{congressbr}
\alias{congressbr}
\alias{congressbr-package}
\title{congressbr.}
\description{
A package for obtaining data from the APIs of the Brazilian
Chamber of Deputies and the Federal Senate, respectively.
}
