% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bill_search.R
\name{sen_bill_search}
\alias{sen_bill_search}
\title{Search for data on legislation in the Brazilian Federal Senate}
\usage{
sen_bill_search(year = "", year_law = "", topic_id = "",
  situation_id = "", date_presented_init = "", date_presented_end = "",
  date_situation_init = "", date_situation_end = "", complementary = "",
  present = "", rapporteur = "", author = "", number = "", type = "",
  law_number = "", keyword = "", party_abbr_author = "",
  author_type = "", law_type = "", author_state = "", in_passage = "",
  ascii = TRUE)
}
\arguments{
\item{year}{\code{integer}. Four-digit year, such as \code{2013}.}

\item{year_law}{\code{integer}. Year of introduction of the law, such as \code{2013}.}

\item{topic_id}{\code{}.}

\item{situation_id}{\code{}.}

\item{date_presented_init}{\code{}.}

\item{date_presented_end}{\code{}.}

\item{date_situation_init}{\code{}.}

\item{date_situation_end}{\code{}.}

\item{complementary}{\code{}.}

\item{present}{\code{}.}

\item{rapporteur}{\code{}.}

\item{author}{\code{}.}

\item{number}{\code{}.}

\item{type}{\code{}.}

\item{law_number}{\code{}.}

\item{keyword}{\code{}.}

\item{party_abbr_author}{\code{}.}

\item{author_type}{\code{}.}

\item{law_type}{\code{}.}

\item{author_state}{\code{}.}

\item{in_passage}{\code{}.}

\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Search for data on legislation in the Brazilian Federal Senate.
}
\examples{
# search for legislation from 2014:
two14 <- sen_bill_search(year = 2014)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
