% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bills.R
\name{sen_bills}
\alias{sen_bills}
\title{Downloads and tidies information on the legislation in the Federal Senate.}
\usage{
sen_bills(type = NULL, number = NULL, year = NULL, ascii = TRUE)
}
\arguments{
\item{type}{\code{character}. The abbreviation of the vote type you're looking
for. A full list of these can be obtained with the \code{sen_bill_list()}
function.}

\item{year}{\code{integer}. Four-digit year, such as \code{2013}.}

\item{ascii}{\code{logical}. If \code{TRUE}, strips Latin characters from
strings.}

\item{number.}{Two-letter abbreviation of Brazilian state. A list of these is
available with the function \code{UF()}.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
pls_5_2010 <- sen_bills(type = "PLS", number = 5, year = 2010)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
