% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bills_various.R
\name{sen_bills_limits}
\alias{sen_bills_limits}
\title{Downloads and tidies information on the types of deadline and time
limits for legislation in the Federal Senate.}
\usage{
sen_bills_limits(ascii = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. If \code{TRUE}, strips Latin characters from
strings.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
sen_bills_limits()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
