% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bills_various.R
\name{sen_bills_situations}
\alias{sen_bills_situations}
\title{Downloads and tidies information on the possible situations a bill
can be in}
\usage{
sen_bills_situations(ascii = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. If \code{TRUE}, strips Latin characters from
strings.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads and tidies information on the possible situations a bill
can be in.
}
\examples{
sen_bills_situations()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
