% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bills_various.R
\name{sen_bills_status}
\alias{sen_bills_status}
\title{Downloads and tidies information on legislation from the current
legislature of the Federal Senate.}
\usage{
sen_bills_status(id = NULL, ascii = TRUE)
}
\arguments{
\item{id}{\code{integer}. This number is the id given to each bill in the
Senate database. For example, running \code{sen_bills_current()} will return a
 dataframe with the variable \code{bill_id} in the first column. These numbers
 can be used as this id.}

\item{ascii}{\code{logical}. If \code{TRUE}, strips Latin characters from
strings.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
sen_bills_status(id = 80406)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
