% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bills_various.R
\name{sen_bills_topics}
\alias{sen_bills_topics}
\title{Downloads and tidies information on the topics of legislation in
the Federal Senate.}
\usage{
sen_bills_topics(active = NULL, ascii = TRUE)
}
\arguments{
\item{active}{\code{character}. Options are "Yes", "No" or \code{NULL}, the
default. "Yes" returns active subtypes, "No" returns inactive subtypes,
while the default returns both.}

\item{ascii}{\code{logical}. If \code{TRUE}, strips Latin characters from
strings.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
top <- sen_bills_topics()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
