% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_budget.R
\name{sen_budget}
\alias{sen_budget}
\title{Downloads and tidies budget information from the Federal Senate}
\usage{
sen_budget(ascii = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads and tidies budget information from the Federal Senate.
}
\examples{
\dontrun{
bud <- sen_budget()
}
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
