% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_plenary.R
\name{sen_plenary_leaderships}
\alias{sen_plenary_leaderships}
\title{Returns information on leaderships in the Federal Senate}
\usage{
sen_plenary_leaderships(parties = TRUE, ascii = TRUE)
}
\arguments{
\item{parties}{\code{logical}. Returns information on party leadership in
the Senate when \code{TRUE}, otherwise returns information on coalitions.}

\item{ascii}{\code{logical}. If \code{TRUE}, the default, strips Latin
characters from the results.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Returns information on leaderships in the Federal Senate.
}
\examples{
party_leaders <- sen_plenary_leaderships()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}

