% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_plenary.R
\name{sen_plenary_sessions}
\alias{sen_plenary_sessions}
\title{Returns the types of sessions in the Federal Senate}
\usage{
sen_plenary_sessions(active = c("Yes", "No"), ascii = TRUE)
}
\arguments{
\item{active}{. Default "Yes", which returns session types which are
still used. Otherwise, returns all types of sessions.}

\item{ascii}{\code{logical}. If \code{TRUE}, the default, strips Latin
characters from the results.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Returns the types of sessions in the Federal Senate.
}
\examples{
sessions <- sen_plenary_sessions()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
