% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_senator.R
\name{sen_senator}
\alias{sen_senator}
\title{Downloads and tidies budget information from the Federal Senate.}
\usage{
sen_senator(code = 0, affiliations = TRUE, mandates = TRUE,
  leaderships = FALSE, absences = FALSE, ascii = TRUE)
}
\arguments{
\item{code}{\code{integer}. Unique code for a senator. A dataframe of these
is available from \code{sen_senator_list()}.}

\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
benedito <- sen_senator(3823) code <- 5529; code <- 3823; code <- 3
### NOTES: check with suplentes
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
