% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_senator_details.R
\name{sen_senator_commissions}
\alias{sen_senator_commissions}
\title{Downloads and tidies information on the commissions on which
senators have served or are serving in the Federal Senate}
\usage{
sen_senator_commissions(id = 0, ascii = TRUE)
}
\arguments{
\item{id}{\code{integer}. This number represents the id of the senator
you wish to get information on. These ids can be extracted from the API
using the \code{sen_senator_list()} function, where they will appear as the
first column in the data frame returned, under the name 'id'.}

\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads and tidies information on the commissions on which
senators have served or are serving in the Federal Senate.
}
\examples{
Armando <- sen_senator_commissions(id = 715)

}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
\seealso{
\code{sen_senator_list()}
}

