% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_senator_specific_details.R
\name{sen_senator_positions}
\alias{sen_senator_positions}
\title{Downloads and tidies specific information related to Senators in
the Federal Senate.}
\usage{
sen_senator_positions(code = 0, active = "both", wide = TRUE,
  list = FALSE)
}
\arguments{
\item{code}{\code{integer}}

\item{active}{\code{character}. Possible values are "yes" or "no". If left
blank, returns all the positions held by the Senator. If "yes", just active
positions, if "no", finished positions.}

\item{wide}{\code{logical}. Default is TRUE, which returns a dataframe of one
row with details on the Senator. When FALSE, returns a two-column dataframe
with the variables (column names when TRUE) in the first column and the values
 in the second. This may be useful in some cases because the wide dataframe
 can have many columns, as the API stores this information in \emph{very}
 deeply nested lists.}

\item{list}{\code{logical}. If TRUE, returns the deeply nested list object
directly from the API. \code{wide} will be set to FALSE automatically
when this option is selected.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
romario <- sen_senator_positions(code = 5322)
head(romario)

}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
