% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_senator_specific_details.R
\name{sen_senator_terms}
\alias{sen_senator_terms}
\title{Get information on the terms served by a senator.}
\usage{
sen_senator_terms(code = 0, wide = TRUE, list = FALSE)
}
\arguments{
\item{code}{\code{integer}. This number represents the code of the senator
you wish to get information on. These codes can be extracted from the API
using the \code{sen_senator_list()} function, where they will appear as the
first column in the data frame returned, under the name 'CodigoParlamentar'.}

\item{wide}{\code{logical}. Default is TRUE, which returns a dataframe of one
row with details on the senator. When FALSE, returns a two-column dataframe
with the variables (column names when TRUE) in the first column and the values
 in the second. This may be useful in some cases because the wide dataframe
 can have many columns, as the API stores this information in \emph{very}
 deeply nested lists.}

\item{list}{\code{logical}. If TRUE, returns the deeply nested list object
directly from the API. \code{wide} will be set to FALSE automatically
when this option is selected.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
lira <- sen_senator_terms(code = 3823)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
