% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_various.R
\name{sen_statement_list}
\alias{sen_statement_list}
\title{Downloads and tidies information on the types of declarations senators
can make in the Federal Senate}
\usage{
sen_statement_list(ascii = TRUE, print = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. TRUE by default, removes latin-1 characters
from returned object.}

\item{print}{\code{logical}. If TRUE (the default), prints the dataframe to
the console.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Downloads and tidies information on the types of declarations senators
can make in the Federal Senate.
}
\examples{
st <- sen_statement_list()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
