% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_votes.R
\name{sen_votes_year_file}
\alias{sen_votes_year_file}
\title{Returns voting information from the Senate floor for the year
requested using xml file method}
\usage{
sen_votes_year_file(year = NULL, binary = TRUE, ascii = TRUE)
}
\arguments{
\item{year}{\code{character} or \code{integer}. Format YYYY}

\item{binary}{\code{logical}. If \code{TRUE}, the default, transforms
votes into \code{1} for "yes", \code{0}, for "no" and \code{NA} for everything
else. If \code{FALSE}, returns a character vector of vote decisions.}

\item{ascii}{\code{logical}. If \code{TRUE}, the default, strips Latin
characters from the results.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Returns voting information from the Senate floor for the year
requested (Only available from 1991 through 2017).
}
\examples{
\donttest{
sen_votes_year_file("2013")
}

}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
