% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statesBR.R
\docType{data}
\name{statesBR}
\alias{statesBR}
\title{A dataframe of Brazilian states, by their \emph{sigla} (acronym) and full
name.}
\format{A data frame with 27 rows and 2 variables}
\usage{
statesBR
}
\description{
\itemize{
  \item UF \code{character}. Acronyms of Brazilian states.
  \item Name \code{character}. Full names of Brazilian states.
}
}
\keyword{datasets}

