% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UF.R
\name{UF}
\alias{UF}
\title{A list of state abbreviations for use with the functions of congressbr.}
\usage{
UF()
}
\description{
This function prints out a character vector of Brazilian
state abbreviations to the console.
}
\note{
This is most useful if you already have an idea of what state you want
 and its abbreviation. If not, to see the full dataset, use \code{data("statesBR")},
 which has the abbreviation, "UF" (\emph{Unidade Federal}) and the full name of the
 states.
}
\author{
Robert Myles McDonnell & Guilherme Jardim Duarte.
}
