% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_bill_search.R
\name{sen_bill_search}
\alias{sen_bill_search}
\title{Search for data on legislation in the Brazilian Federal Senate}
\usage{
sen_bill_search(year = "", year_law = "", topic_id = "",
  situation_id = "", date_presented_init = "", date_presented_end = "",
  date_situation_init = "", date_situation_end = "", complementary = "",
  present = "", rapporteur = "", author = "", number = "", type = "",
  law_number = "", keyword = "", party_abbr_author = "",
  author_type = "", law_type = "", author_state = "", in_passage = "",
  ascii = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
Search for data on legislation in the Brazilian Federal Senate.
}
\examples{
# search for legislation from 2014:
two14 <- sen_bill_search(year = 2014)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
