% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_commission_list.R
\name{sen_commissions}
\alias{sen_commissions}
\alias{sen_commissions}
\title{Information on commissions in the Federal Senate.}
\usage{
sen_commissions(type = c("permanent", "cpi", "temporary"), ascii = TRUE)

sen_commissions(type = c("permanent", "cpi", "temporary"), ascii = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}

\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.

A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\details{
Returns a data frame with the following variables:
\itemize{
 \item{\code{commission_id: }}{unique code for each commission.}
 \item{\code{commission_abbr: }}{.}
 \item{\code{commission__name: }}{.}
}

Returns a data frame with the following variables:
\itemize{
 \item{\code{commission_id: }}{unique code for each commission.}
 \item{\code{commission_abbr: }}{.}
 \item{\code{commission__name: }}{.}
}
}
\examples{
jobs <- sen_commissions()
jobs <- sen_commissions()
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.

Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
