% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_part_dec_comm.R
\name{sen_parties}
\alias{sen_parties}
\title{Downloads and tidies information on the political parties in the
Federal Senate.}
\usage{
sen_parties(ascii = TRUE)
}
\arguments{
\item{ascii}{\code{logical}. TRUE by default, removes latin-1 characters
from returned object.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
\\code{party <- sen_parties(); head(party)}
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
