% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_senator_details.R
\name{sen_senator_details}
\alias{sen_senator_details}
\title{Downloads and tidies information on the Senators in the Federal Senate.}
\usage{
sen_senator_details(code = 0)
}
\arguments{
\item{code}{(\code{integer}). This number represents the code of the senator
you wish to get information on. These codes can be extracted from the API
using the \code{sen_senator_list()} function, where they will appear as the
first column in the data frame returned, under the name 'CodigoParlamentar'.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
Acir_G <- sen_senator_details(code = 4981)

}
\seealso{
sen_senator_list
}
\author{
Robert Myles McDonnell & Guilherme Jardim Duarte.
}
