% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_senator_details.R
\name{sen_senator_mandates}
\alias{sen_senator_mandates}
\title{Downloads and tidies information on senators' mandates in the Federal
Senate.}
\usage{
sen_senator_mandates(code = 0, ascii = TRUE)
}
\arguments{
\item{code}{\code{integer}. This number represents the code/id of the senator
you wish to get information on. These codes can be extracted from the API
using the \code{sen_senator_list()} function, where they will appear as the
first column in the data frame returned, under the name 'id'.}

\item{ascii}{\code{logical}. If TRUE, certain strings are converted to ascii
format.}
}
\value{
A tibble, of classes \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\examples{
terms <- sen_senator_mandates(code = 4763)
terms <- sen_senator_mandates(code = 3398)
}
\seealso{
\code{sen_senator_list()}
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
