% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_parties_declarations.R
\name{sen_statement_list}
\alias{sen_statement_list}
\title{Downloads and tidies information on the types of declarations senators
can make in the Federal Senate.}
\usage{
\code{sen_statement_list(return = TRUE, ascii = TRUE)}
}
\arguments{
\item{return}{\code{logical}. If TRUE, returns a data frame (tibble).}

\item{ascii}{\code{logical}. TRUE by default, removes latin-1 characters
from returned object.}
}
\value{
The data frame is printed to the console. If \code{return = TRUE} is
chosen, the data frame is printed to the console and also returned as an object.
}
\examples{
st <- sen_statement_list(return = T)
}
\author{
Robert Myles McDonnell, Guilherme Jardim Duarte & Danilo Freire.
}
